Kyle feedback
