+++ kubedog.go
@@ -76,8 +76,7 @@ func (kdt *Test) Run() {
 	kdt.scenarioContext.Step(`^an Auto Scaling Group named ([^"]*)$`, kdt.AwsContext.AnASGNamed)
 	kdt.scenarioContext.Step(`^(?:I )?update (?:the )?current Auto Scaling Group with ([^"]*) set to ([^"]*)$`, kdt.AwsContext.UpdateFieldOfCurrentASG)
 	kdt.scenarioContext.Step(`^(?:the )?current Auto Scaling Group (?:is )?scaled to \(min, max\) = \((\d+), (\d+)\)$`, kdt.AwsContext.ScaleCurrentASG)
-	kdt.scenarioContext.Step(`^(?:the )?DNS name (\S+) (?:should be )?created in hostedZoneID (\S+)$`, kdt.AwsContext.DnsNameInHostedZoneID)
-	kdt.scenarioContext.Step(`^(?:the )?DNS name (\S+) (?:should )?not (?:be )?created in hostedZoneID (\S+)$`, kdt.AwsContext.DnsNameNotInHostedZoneID)
+	kdt.scenarioContext.Step(`^(?:the )?DNS name (\S+) (should|should not) be created in hostedZoneID (\S+)$`, kdt.AwsContext.DnsNameShouldOrNotInHostedZoneID)
 	kdt.scenarioContext.Step(`^(?:I )?(add|remove) (?:the )?(\S+) role as trusted entity to iam role ([^"]*)$`, kdt.AwsContext.IamRoleTrust)
 	kdt.scenarioContext.Step(`^(?:I )?(add|remove) ?([^"]*) as trusted entity to iam role ([^"]*)$`, kdt.AwsContext.IamRoleTrust)
 	kdt.scenarioContext.Step(`^(?:I )?(add|remove) cluster shared iam role$`, kdt.AwsContext.ClusterSharedIamOperation)
